Implent encoding of create syscall
I could only trigger the ES_DESTINATION_TYPE_EXISTING_FILE case,
the ES_DESTINATION_TYPE_NEW_PATH case probably only triggers for AUTH events.

Decoding the acl was a bit hairy since the acl in the message needs to be converted first.
If we just call acl_to_text() on it we'll crash in acl_get_entry in libc.